+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added syntax rule for `None` literal
+
+### Fixed
+- Profile header rule now strictly requires full semver version in version field, as the spec defines
+- Condition to add SingleLineCommentTrivia to termina tokens in sublexer
 
 ## [2.0.0] - 2022-11-08
 ### Added